:recycle:Refactor: GridSection 경로 수정
