+++ src/pages/Error/LoginRequired.jsx
@@ -1,8 +1,8 @@
 import { useNavigate } from 'react-router';
 
-import GridSection from '@/components/background/GridSection';
 import RedirectButton from '@/components/common/Button/RedirectButton';
 import StatusMessage from '@/components/common/StatusMessage';
+import GridSection from '@/components/layout/background/GridSection';
 
 export default function LoginRequired() {
   const title = '로그인이 필요한 서비스입니다.';
